+++ corebehrt/modules/cohort_handling/patient_filter.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 import pandas as pd
 
-from corebehrt.modules.setup.constants import (
+from corebehrt.functional.constants import (
     PID_COL,
     TIMESTAMP_COL,
     BIRTHDATE_COL,
