+++ corebehrt/modules/setup/loader.py
@@ -1,29 +1,14 @@
 import logging
 from os.path import join
-from typing import Tuple
 
 import torch
 from transformers import ModernBertConfig
 
-from corebehrt.modules.setup.config import Config, load_config
 from corebehrt.functional.setup.model import get_last_checkpoint_epoch
 
 logger = logging.getLogger(__name__)  # Get the logger for this module
 
-VOCABULARY_FILE = "vocabulary.pt"
 CHECKPOINT_FOLDER = "checkpoints"
-
-
-def load_checkpoint_and_epoch(model_dir: str, checkpoint_epoch: str = None) -> Tuple:
-    """Load checkpoint and epoch from config."""
-    checkpoint = ModelLoader(
-        model_dir, checkpoint_epoch=checkpoint_epoch
-    ).load_checkpoint()
-    if checkpoint is not None:
-        epoch = checkpoint["epoch"]
-    else:
-        epoch = get_last_checkpoint_epoch(join(model_dir, CHECKPOINT_FOLDER))
-    return checkpoint, epoch
 
 
 class ModelLoader:
